pyproject.toml
[tool.poetry]
name = "local-airflow"
version = "0.1.0"
description = ""
authors = ["gagocaan <gagocaan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "^2.6.2"
hydra-core = "^1.3.2"
loguru = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
